Added the incorrect save dir handling

+++ console/Output/FileOutput.cs
@@ -1,3 +1,4 @@
+using System;
 using System.IO;
 using console.Dto;
 
@@ -24,13 +25,23 @@ namespace console.Output
             _applicationConfig.ReportOutputPath = outputPath;
         }
 
-        public void saveReport(ReportDownloadTask task)
+        public bool saveReport(ReportDownloadTask task)
         {
             var downloadedReportFileName = Path.Combine(this._applicationConfig.BrowserDownloadsDirectory, "Документ_об_образовании.pdf");
+            if (!File.Exists(downloadedReportFileName))
+            {
+                Console.WriteLine("Не удалось найти загруженный файл " + downloadedReportFileName);
+                Console.WriteLine("Проверьте второй аргумент запускной команды, отвечающий за путь до директории для файлов, загружаемых Mozilla Firefox (по умолчанию: C:\\Users\\<Имя пользователя>\\Downloads>)");
+                return false;
+            }
+
+            
+            
             File.Copy(
                 downloadedReportFileName, 
                 Path.Combine(_applicationConfig.ReportOutputPath, getReportFileName(task))
             );
+            return true;
         }
 
         public bool isReportExists(ReportDownloadTask task)
